requirements.txt
Flask
spotdl
ffmpeg
python-dotenv
azure-identity
azure-storage-blob